Application/requirements.txt
-i https://pypi.org/simple
applicationinsights==0.11.7
atomicwrites==1.3.0
attrs==18.2.0
bcrypt==3.1.6
blueprints==2.3.0.2
cffi==1.11.5
click==7.0
flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
jinja2==2.10
markupsafe==1.1.0
more-itertools==5.0.0
pendulum==2.0.4
pluggy==0.8.1
py==1.7.0
pycparser==2.19
pytest==4.2.0
python-dateutil==2.8.0
pytzdata==2018.9
six==1.12.0
werkzeug==0.14.1
